requirements.txt
django>2.2.0,<=2.2.10
djangorestframework==3.9.2
